freeze_requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.0
altair==5.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.1.1
cuda-python==12.2.0
cycler==0.12.1
Cython==3.0.9
distlib==0.3.8
dlib==19.24.2
docker-pycreds==0.4.0
dotmap==1.3.30
EasyProcess==1.1
einops==0.3.0
entrypoint2==1.1
facenet-pytorch==2.6.0
filelock==3.13.1
fire==0.6.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
glfw==2.7.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
huggingface-hub==0.24.0
identify==2.5.35
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==7.0.1
importlib-resources==6.1.2
imutils==0.5.4
invisible-watermark==0.2.0
iopath==0.1.10
jax==0.4.13
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.0
lazy-loader==0.3
lightning-lite==1.8.0
lightning-utilities==0.11.5
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mediapipe==0.10.10
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
nodeenv==1.8.0
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.16.2
onnx2torch==1.5.15
open-clip-torch==2.0.2
opencv-contrib-python==4.9.0.80
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
patool==1.12
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
plotly==5.19.0
portalocker==2.8.2
pre-commit==3.5.0
protobuf==3.20.3
psutil==5.9.8
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pyDeprecate==0.3.1
pygments==2.17.2
Pympler==1.0.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytorch-msssim==1.0.0
-e git+https://github.com/facebookresearch/pytorch3d.git@38cf0dc1c52138987e6e66295c5a2d192a6914bd#egg=pytorch3d
pytz==2024.1
pyunpack==0.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
semver==3.0.2
sentry-sdk==1.42.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
test-tube==0.7.5
threadpoolctl==3.3.0
tifffile==2023.7.10
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torcheval==0.0.7
torchmetrics==0.6.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
transformers==4.42.0
triton==2.2.0
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.1
wandb==0.16.4
watchdog==4.0.0
wcwidth==0.2.13
webdataset==0.2.5
werkzeug==3.0.1
xformers==0.0.24
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0